watermark parameters
