+++ server/src/main/java/org/apache/iotdb/db/conf/IoTDBConfig.java
@@ -23,6 +23,8 @@ import java.time.ZoneId;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 import org.apache.iotdb.db.metadata.MManager;
 import org.apache.iotdb.db.service.TSServiceImpl;
 import org.slf4j.Logger;
@@ -35,6 +37,9 @@ public class IoTDBConfig {
   private static final String MULTI_DIR_STRATEGY_PREFIX =
       "org.apache.iotdb.db.conf.directories.strategy.";
   private static final String DEFAULT_MULTI_DIR_STRATEGY = "MaxDiskUsableSpaceFirstStrategy";
+
+  /* Names of Watermark methods */
+  public static final String WATERMARK_GROUPED_LSB = "GroupBasedLSBMethod";
 
   private String rpcAddress = "0.0.0.0";
 
@@ -243,6 +248,8 @@ public class IoTDBConfig {
    * Bit string of watermark
    */
   private String watermarkBitString = "11001010010101";
+
+  private String watermarkMethod = "GroupBasedLSBMethod(mark_rate=5,max_right_bit=5)";
 
   public IoTDBConfig() {
     // empty constructor
@@ -657,4 +664,35 @@ public class IoTDBConfig {
   public void setWatermarkBitString(String watermarkBitString) {
     this.watermarkBitString = watermarkBitString;
   }
+
+  public String getWatermarkMethod() {
+    return watermarkMethod.split("\\(")[0];
+  }
+
+  public int getWatermarkParamMarkRate() {
+    return Integer.parseInt(getWatermarkParamValue("mark_rate", "5"));
+  }
+
+  public int getWatermarkParamMaxRightBit() {
+    return Integer.parseInt(getWatermarkParamValue("max_right_bit", "5"));
+  }
+
+  public String getWatermarkParamValue(String key, String defaultValue) {
+    String res = getWatermarkParamValue(key);
+    if (res != null) {
+      return res;
+    }
+    return defaultValue;
+  }
+
+  public String getWatermarkParamValue(String key) {
+    String pattern = key + "=(\\w*)";
+    Pattern r = Pattern.compile(pattern);
+    Matcher m = r.matcher(watermarkMethod);
+    if (m.find() && m.groupCount() > 1) {
+      return m.group(1);
+    }
+    return null;
+  }
+
 }
