+++ server/src/main/java/org/apache/iotdb/db/utils/QueryDataSetUtils.java
@@ -19,6 +19,7 @@ import java.nio.ByteBuffer;
 import java.util.ArrayList;
 import java.util.List;
 import org.apache.iotdb.db.conf.IoTDBConfig;
+import org.apache.iotdb.db.tools.GroupedLSBWatermarkEncoder;
 import org.apache.iotdb.db.tools.WatermarkEncoder;
 import org.apache.iotdb.service.rpc.thrift.TSDataValue;
 import org.apache.iotdb.service.rpc.thrift.TSQueryDataSet;
@@ -53,8 +54,14 @@ public class QueryDataSetUtils {
     tsQueryDataSet.setRecords(new ArrayList<>());
     WatermarkEncoder watermarkEncoder = null;
     if (conf != null && conf.isEnableWatermark()) {
-      watermarkEncoder = new WatermarkEncoder(conf.getWatermarkSecretKey(),
-          conf.getWatermarkBitString());
+      switch (conf.getWatermarkMethod()) {
+        case IoTDBConfig.WATERMARK_GROUPED_LSB:
+          watermarkEncoder = new GroupedLSBWatermarkEncoder(conf);
+          break;
+        default:
+          throw new UnSupportedDataTypeException(String.format(
+              "Watermark method is not supported yet: %s", conf.getWatermarkMethod()));
+      }
     }
     for (int i = 0; i < fetchSize; i++) {
       if (queryDataSet.hasNext()) {
